invoice-auditor/requirements.txt
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langgraph>=0.2.29
langchain-google-genai>=2.0.1
langchain-huggingface>=0.0.1
langgraph.checkpoint.sqlite>=0.1.0
sentence-transformers>=2.2.0

PyPDF2>=3.0.1
pdfminer.six>=20231228
python-docx>=1.0.1
pillow>=10.0.0
pytesseract>=0.3.10

pyyaml>=6.0.1
streamlit>=1.37.0
fastapi>=0.115.0
uvicorn>=0.30.0
chromadb>=0.4.22
langchain-chroma>=0.1.2
weasyprint>=60.0
jinja2>=3.1.0
jq
loguru>=0.7.0